Unit Test Skeletons
It was very hard to write unit tests without validations, so I began writing/pondering the validation code. This will need to be integrated into the React and potentially Django code. I will be asking several questions this week. For the scope of the project, manual testing may be faster.